skeleton of CLI
