+++ agent/provider/src/cli.rs
@@ -3,9 +3,9 @@ pub mod clean;
 pub mod config;
 pub mod exe_unit;
 pub mod keystore;
-pub mod outbound_config;
 pub mod preset;
 pub mod profile;
+pub mod rule;
 pub mod whitelist;
 
 use crate::startup_config::ProviderConfig;
